# apps/docs/openapi/writing-app-api.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Writing App API",
    "description": "개발 문서",
    "version": "0.0.1"
  },
  "paths": {
...
    "/profile": {
      "get": {
        "operationId": "getProfile",
        "responses": {
          "200": {
            "description": "현재 학습자의 프로필 요약입니다.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "courseCount": {
                      "type": "integer",
                      "minimum": 0,
                      "maximum": 9007199254740991
                    },
                    "completedLessonCount": {
                      "type": "integer",
                      "minimum": 0,
                      "maximum": 9007199254740991
                    }
                  },
                  "required": ["courseCount", "completedLessonCount"]
                }
              }
            }
          },
          "401": {
            "description": "로그인이 필요합니다.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string",
                      "const": "unauthorized"
                    },
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": ["code", "message"]
                }
              }
            }
          },
          "503": {
            "description": "데이터베이스를 사용할 수 없습니다.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string",
                      "const": "database-unavailable"
                    },
                    "message": {
                      "type": "string",
                      "const": "데이터베이스를 사용할 수 없습니다."
                    }
                  },
                  "required": ["code", "message"]
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}